MEDAI_BACKEND/requirements.txt
accelerate==1.12.0
annotated-doc==0.0.4
anyio==4.12.1
asgiref==3.11.1
astroid==4.0.4
audioread==3.1.0
bitsandbytes==0.49.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
decorator==5.2.1
dill==0.4.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
filelock==3.20.0
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
isort==7.0.0
Jinja2==3.1.6
jiwer==4.0.0
joblib==1.5.3
lazy_loader==0.4
Levenshtein==0.27.3
librosa==0.11.0
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
packaging==26.0
pillow==12.1.1
platformdirs==4.9.0
pooch==1.9.0
psutil==7.2.2
pycparser==3.0
Pygments==2.19.2
PyYAML==6.0.3
RapidFuzz==3.14.3
regex==2026.1.15
requests==2.32.5
rich==14.3.2
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
setuptools==70.2.0
shellingham==1.5.4
soundfile==0.13.1
soxr==1.0.0
sqlparse==0.5.5
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
tomlkit==0.14.0
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.3
transformers==5.1.0
typer==0.23.1
typer-slim==0.23.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
